Accept sub-curator role ui update (#202)
