+++ src/components/curator-actions/AcceptCuratorRole.svelte
@@ -1,30 +1,31 @@
 <script lang="ts">
-	import { convertPlanckToDot, dryRunAndSubmitTransaction } from '../../../utils/polkadot';
-	import BountyDialog from '../../BountyDialog.svelte';
+	import type { Bounty } from '../../types/bounty';
+	import { convertPlanckToDot, dryRunAndSubmitTransaction } from '../../utils/polkadot';
+	import BountyDialog from '../BountyDialog.svelte';
 	import { ApiRx, WsProvider } from '@polkadot/api';
 	import { firstValueFrom } from 'rxjs';
-	import { activeAccount } from '../../../stores';
+	import { activeAccount } from '../../stores';
 	import { onMount } from 'svelte';
 	import {
 		showErrorDialog,
 		showLoadingDialog,
 		showSuccessDialog
-	} from '../../../utils/loading-screen';
-	import ToggleIcon from '../../svg/ToggleIcon.svelte';
-	import type { ChildBounty } from '../../../types/child-bounty';
-	import { WALLET_CONNECT_SOURCE } from '../../../utils/WcSigner';
+	} from '../../utils/loading-screen';
+	import ToggleIcon from '../svg/ToggleIcon.svelte';
+	import { WALLET_CONNECT_SOURCE } from '../../utils/WcSigner';
 
 	export let open = false;
-	export let childBounty: ChildBounty;
+	export let bounty: Bounty;
 
 	let fee = '-';
 	let isToggled = false;
+	export { className as class };
 
 	onMount(async () => {
 		await calculateFee();
 	});
 
-	async function acceptCuratorRule() {
+	async function acceptCuratorRole() {
 		open = false;
 		showLoadingDialog('Submitting transaction');
 		try {
@@ -35,10 +36,7 @@
 			const wsProvider = new WsProvider('ws://localhost:8000');
 			const api = await firstValueFrom(ApiRx.create({ provider: wsProvider }));
 
-			let transaction = api.tx.childBounties.acceptCurator(
-				childBounty.parentBounty,
-				childBounty.id
-			);
+			let transaction = api.tx.bounties.acceptCurator(bounty.id);
 			const { errorMessage, result } = await dryRunAndSubmitTransaction(
 				api,
 				transaction,
@@ -75,10 +73,7 @@
 		try {
 			const wsProvider = new WsProvider('ws://localhost:8000');
 			const api = await firstValueFrom(ApiRx.create({ provider: wsProvider }));
-			let transaction = api.tx.childBounties.acceptCurator(
-				childBounty.parentBounty,
-				childBounty.id
-			);
+			let transaction = api.tx.bounties.acceptCurator(bounty.id);
 
 			let observableFee = transaction.paymentInfo($activeAccount.address);
 			fee =
@@ -94,9 +89,9 @@
 <BountyDialog bind:open title="ACCEPT CURATOR ROLE">
 	<section class="space-y-5">
 		<div class="space-x-1">
-			<span>#{childBounty.id}</span>
-			{#if childBounty.description !== undefined}
-				<span>{childBounty.description}</span>
+			<span>#{bounty.id}</span>
+			{#if bounty.description !== undefined}
+				<span>{bounty.description}</span>
 			{/if}
 		</div>
 
@@ -121,9 +116,9 @@
 	</section>
 
 	<button
-		on:click={acceptCuratorRule}
+		on:click={acceptCuratorRole}
 		disabled={!isToggled}
-		class="{`w-full md:w-fit mt-10 ${isToggled ? 'button-popup' : 'opacity-50 cursor-not-allowed'}`}
+		class="{`w-full md:w-fit mt-10 h-12 ${isToggled ? 'button-popup' : 'opacity-50 cursor-not-allowed'}`}
   {`${!isToggled ? 'button-popup' : 'cursor-allowed'}`}">SIGN</button
 	>
 </BountyDialog>
